+++ TagsCloudContainer/Drawer/RectangleLayout.cs
@@ -11,32 +11,32 @@ namespace TagsCloudContainer.Drawer
         private readonly IOptions options;
         private readonly ILayouter layouter;
         private readonly ILayoutDrawer drawer;
-        private readonly List<WordRectangle> rectangles;
+        private readonly Bitmap bitmap;
+        private readonly Graphics graphics;
 
         public RectangleLayout(ILayouter layouter, ILayoutDrawer drawer, IOptions options)
         {
             this.layouter = layouter;
             this.drawer = drawer;
             this.options = options;
-            rectangles = new List<WordRectangle>();
+            bitmap = new Bitmap(options.Width, options.Height);
+            graphics = Graphics.FromImage(bitmap);
         }
 
         public void PlaceWords(Dictionary<string, int> words)
         {
-            layouter.SetCenter(options.Width / 2, options.Height / 2);
+            layouter.SetCenter(new Point(options.Width / 2, options.Height / 2));
             foreach (var (word, count) in words)
             {
                 var fontSize = CalculateFontSize(count);
                 var rectangle = layouter.PutNextRectangle(GetWordSize(word, fontSize, options.FontFamily));
-                rectangles.Add(new WordRectangle(rectangle, word, fontSize));
+                drawer.AddRectangle(new WordRectangle(rectangle, word, fontSize));
             }
-
-            drawer.AddRectangles(rectangles);
         }
 
         private Size GetWordSize(string word, int fontSize, string fontFamily)
         {
-            return drawer.Graphics.MeasureString(word, new Font(fontFamily, fontSize)).ToSize();
+            return graphics.MeasureString(word, new Font(fontFamily, fontSize)).ToSize();
         }
 
         private static int CalculateFontSize(int wordCount)
@@ -46,14 +46,14 @@ namespace TagsCloudContainer.Drawer
 
         public void DrawLayout()
         {
-            drawer.Draw();
+            drawer.Draw(graphics);
         }
 
         public void SaveLayout()
         {
             var outputDirectory = options.OutputDirectory ?? Directory.GetCurrentDirectory();
             var fullPath = Path.Combine(outputDirectory, options.OutputFileName + options.OutputFileExtension);
-            drawer.Bitmap.Save(fullPath);
+            bitmap.Save(fullPath);
             Console.WriteLine($"Tag cloud visualization saved to file {fullPath}");
         }
     }
