+++ TagsCloudContainer/WordsParser/IFiltercs.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace TagsCloudContainer.WordsParser
+{
+    public interface ISettings
+    {
+        public IEnumerable<string> BoringWords { get; }
+    }
+}
