--- TagsCloudContainer/WordsParser/ISettings.cs
@@ -1,7 +1,0 @@
-﻿namespace TagsCloudContainer.WordsParser
-{
-    public interface ISettings
-    {
-        public bool ToInitialForm { get; }
-    }
-}
